alterações do projeto

+++ Segundo_Modulo/LinguagemProgramacao_II/ProjetosAvaliativos/TI22_LP2_MairyEduardo_ExercicioAvaliativo01/src/main/java/br/com/org/fundatec/model/Veiculo.java
@@ -5,7 +5,7 @@ public abstract class Veiculo {
     private String marca;
     private String modelo;
 
-    abstract void calculoTarifa(int tempo);
+    abstract void calculoTarifa(Double tempo);
 
     public Veiculo (String marca, String modelo){
         this.marca = marca;
